requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
flask>=1.0.2
flask-cors>=3.0.7
gunicorn==19.5.0
Jinja2>=2.8
Werkzeug>=0.10.4
beautifulsoup4==4.5.3
requests>=2.20.0
coverage==4.1
nose==1.3.7
setupext-janitor==1.0.0
bs4==0.0.1
Pillow==3.0
